# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ISP Billing API
  version: 1.0.0
servers:
  - url: https://your-host/api
paths:
...
  /team/members:
    get:
      summary: List tenant team members and their roles
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '403': { description: Owner or admin role required }
  /team/members/{id}/role:
    patch:
      summary: Update a tenant team member role
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '403': { description: Insufficient role }
        '404': { description: Team member not found }
        '409': { description: Last owner protection }
  /team/members/{id}:
    delete:
      summary: Remove a team member from the tenant
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '403': { description: Insufficient role }
        '404': { description: Team member not found }
        '409': { description: Last owner or self-removal protection }
  /invites:
    get:
      summary: List pending tenant invites
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '403': { description: Owner or admin role required }
    post:
      summary: Create a tenant invite for owner, admin, or operator access
      security:
        - bearerAuth: []
      responses:
        '201': { description: Created }
        '400': { description: Invalid payload }
        '403': { description: Insufficient role for requested invite role }
        '409': { description: User is already a member }
  /invites/accept:
    post:
      summary: Publicly accept a tenant invite and create a tenant session
      responses:
        '200': { description: OK }
        '400': { description: Invalid or expired invite }
  /invites/{id}:
    delete:
      summary: Revoke a pending tenant invite
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '403': { description: Insufficient role }
        '404': { description: Invite not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY